Carry the erosion parameters at the width the solver computes in
GraphHydraulicAccelerationGate [A2] measured 0.00079 m of drift over 15
iterations against a 0.0002 m budget, while [A1] measured exactly zero over
one. The kernel and the oracle are arithmetic twins — both compute in double
and round only where they write a grid — so an exact first pass could not
become a drifting fifteenth on its own.

It was the parameters. ErosionHydraulicParams stored them as float, so 0.05,
0.02, 0.4 and 0.01 arrived about 1e-9 away from the doubles the oracle read
out of the same Dictionary. One pass still agreed bit-for-bit; by the second
the water grid was a float32 ULP apart, and `sed_c < cap` turns a ULP into a
whole erode-or-deposit decision for that cell. From there it doubled every few
iterations: 39 cells at pass 2, 2010 and 8e-4 m by pass 15.

The fields are now double, which is the width the solver already used for
every intermediate. All 15 iterations are bit-identical. The graph path is
untouched — its parameters come from a float32 program, and a float widened to
double is the same number — and the GPU push constants cast back to float at
the boundary where they always did.

[A2]'s budget was sized for that drift, so it now shares [A1]'s bit-level one,
with a control proving a 1e-4 m error in a single cell is still detected at
that width.

GraphGpuParity (windowed), GraphCppParity, GraphSolverFreeze, GraphDevTwin,
GraphNodeCaching, GraphFold, GraphWorkerThread, GraphHydraulicParticle,
GraphHydraulicStreamLog and BrushStack all pass.

+++ src/pasture_3d_erosion_hydraulic.cpp
@@ -15,22 +15,22 @@ ErosionHydraulicParams ErosionHydraulicParams::from_dict(const Dictionary &p_dic
 		p.iterations = std::max(1, (int)p_dict["iterations"]);
 	}
 	if (p_dict.has("rain_rate")) {
-		p.rain_rate = std::max(0.0f, (float)p_dict["rain_rate"]);
+		p.rain_rate = std::max(0.0, (double)p_dict["rain_rate"]);
 	}
 	if (p_dict.has("evaporation_rate")) {
-		p.evaporation_rate = std::clamp((float)p_dict["evaporation_rate"], 0.0f, 1.0f);
+		p.evaporation_rate = std::clamp((double)p_dict["evaporation_rate"], 0.0, 1.0);
 	}
 	if (p_dict.has("sediment_capacity")) {
-		p.sediment_capacity = std::max(0.0f, (float)p_dict["sediment_capacity"]);
+		p.sediment_capacity = std::max(0.0, (double)p_dict["sediment_capacity"]);
 	}
 	if (p_dict.has("erosion_speed")) {
-		p.erosion_speed = std::clamp((float)p_dict["erosion_speed"], 0.0f, 1.0f);
+		p.erosion_speed = std::clamp((double)p_dict["erosion_speed"], 0.0, 1.0);
 	}
 	if (p_dict.has("deposition_speed")) {
-		p.deposition_speed = std::clamp((float)p_dict["deposition_speed"], 0.0f, 1.0f);
+		p.deposition_speed = std::clamp((double)p_dict["deposition_speed"], 0.0, 1.0);
 	}
 	if (p_dict.has("min_slope")) {
-		p.min_slope = std::max(0.0f, (float)p_dict["min_slope"]);
+		p.min_slope = std::max(0.0, (double)p_dict["min_slope"]);
 	}
 	return p;
 }
